Firestore Integration Checkpoint
