+++ functions/src/firebase.ts
@@ -1,27 +1,11 @@
-import {onRequest} from "firebase-functions/https";
-import {logger} from "firebase-functions/v1";
-import {db} from "./fire";
+/**
+ * Initialize Firebase Admin
+ * This enables the initialization of various clients (i.e., Firestore)
+ */
+import * as admin from "firebase-admin";
 
+// Initialize Firebase Admin
+admin.initializeApp();
 
-// Create a new document in Firestore
-export const getAccountQuota = onRequest(async (request, response) => {
-  try {
-    logger.info("Retrieving Account Quota");
-
-    db.collection("account").doc("requests").get()
-      .then((doc) => {
-        if (doc.exists) {
-          logger.info("Document data:", doc.data());
-          response.json({"accountQuota": doc.data()});
-        } else {
-          logger.warn("No such document!");
-        }
-      })
-      .catch((error) => {
-        logger.error("Error getting document:", error);
-      });
-  } catch (error) {
-    logger.error("Error creating user:", error);
-    response.status(500).send("Internal Server Error");
-  }
-});
+// Firestore Client
+export const db = admin.firestore();
